map building simulation script
This script simulates the map building mechanics that I had planned,
with measurements for minimum traversal time of the map.

The simulations showed that I need to ditch the idea of shuffling the
clue encounters into the location encounter decks. For any number of
encounters worth being called a "deck" the average traversal time of the
map is more turns than desirable.

In retrospect that makes some sense because that mechanic was primarily
borrowed from the Pathfinder Adventure Card Game where digging through
the decks within the time limit was one of the primary obstacles in the
game.

In this game, though, the clue encounters are only a sub-obstacle to
moving onward towards the next phase of the game.

Getting rid of that mechanic seems to free up the design in a couple of
ways.

First, I had been concerned about the setup and teardown time required
to create and disperse the encounter decks. With those decks no longer
needed to hide clue encounters the decks can be replaced with something
else, such as an encounter table, that doesn't require setup and
teardown.

Second, with that obstacle removed it feels (at least from a design
perspective) that there is room for other requirements and/or
more-complex (and varied) clue encounters - or location-specific
requirements to get to them.