Revised dice sim for better utility
Reworked internal structure of the simulation script to be more
extensible in interesting directions.

The previous version made it easy to mix and match elemental dice and
build attributes, but that's actually a step beyond where I'm at in this
stage of exploratory prototyping.

The new version is better suited to playing around with the composition
of the dice - what (and how many) symbols are on each face.

It also adds some additional commentary on how/why different dice with
the same EV(Warp), + 1/6, can be more or less interesting.