Expand map sim capabilities
One of the potential balancing obstacles is the long tail that's
possible for the length of the clue trail based on the number of
possible clue locations. This was theoretically (L-(C-1)) where L is the
number of locations that can have clues and C is the number of clues.

For example, with 15 clue locations (the number of spaces in base AH2
that can receive clue tokens) and 3 clues, the worst-case is that two of
the clues are found immediately and the third clue has to be chased
through all 13 of the remaining locations, with a mean number of
locations that have to be searched per clue of ~3.99.  This would mean
that in the average case the team would be forced to visit 12 locations
to track down all of the clues.

The idea is to force the creation of additional dead-ends by setting
aside additional clue targets as if they were going to be their own clue
trails but then leaving them unused.

Taking the prior example, if an additional 3 locations are set aside
(ideally unrevealed) as unusued clues that forces the worst-case down to
10 locations and pushes the mean number of locations per clue down to
~2.29.  This pushes the number of locations that must be visited in the
average case down to 7.

This feels like it will be a critical lever for tuning the pacing of the
game.