Update success counting logic for warping dice
While looking at the odds of success for various numbers of dice and
various target numbers for successes it felt as if the ways in which the
difficulty of checks could be tuned was insufficient, especially if
penalties to the number of dice are prohibited or discouraged.

This new method gives the possiblity of more than one success per die in
all but adverse (depowered) circumstances. This effectively increases
the range of the target number of successes, giving more tuning options.