Update commentary in dice_sim.py
Add reference points for attribute scores (i.e. the number of dice that
will be rolled in checks).

Removed speculative philosophizing about the effect of increased
variance around the same midpoint.

Added discussion around the change to counting successes.

Replaced hand-calculated odds at various numbers of dice with a more
detailed preview of the range of results for a single die.

Added discussion around the use of penalties to attribute scores as part
of tuning the difficulty of checks.

Tweaked wording in a few other places.