Refinement of dice symbology and result counting
Among other things, this reverts d7c4a00595, making each symbol once
more count as only one in the results.

This accompanies changes to the face composition so that most faces now
contain more than one symbol, as I think this is likely to have lower
cognitive load.

This is also important because I've merged the so-called "mid" and "mad"
warping dice into the "regular" warping die, also based on the power
context, but with a desired distribution that can't be met with the
logic used in d7c4a00595 for success symbols.

Having different rules for Warp/Ward from the rules for Success would
definitely be too much cognitive load.

This also adds functionality to the tallying of results to allow for
static Ward and an option to disregard excess Ward.